--- src/main/java/com/sparta/currency_user/currency/controller/CurrencyController.java
@@ -1,32 +1,0 @@
-package com.sparta.currency_user.currency.controller;
-
-import com.sparta.currency_user.currency.dto.CurrencyRequestDto;
-import com.sparta.currency_user.currency.dto.CurrencyResponseDto;
-import com.sparta.currency_user.currency.service.CurrencyService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/currencies")
-@RequiredArgsConstructor
-public class CurrencyController {
-    private final CurrencyService currencyService;
-
-    @GetMapping
-    public ResponseEntity<List<CurrencyResponseDto>> findCurrencies() {
-        return ResponseEntity.ok().body(currencyService.findAll());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<CurrencyResponseDto> findCurrency(@PathVariable Long id) {
-        return ResponseEntity.ok().body(currencyService.findById(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<CurrencyResponseDto> createCurrency(@RequestBody CurrencyRequestDto currencyRequestDto) {
-        return ResponseEntity.ok().body(currencyService.save(currencyRequestDto));
-    }
-}
