--- src/main/java/com/sparta/currency_user/currency/dto/CurrencyRequestDto.java
@@ -1,21 +1,0 @@
-package com.sparta.currency_user.currency.dto;
-
-import com.sparta.currency_user.currency.entity.Currency;
-import lombok.Getter;
-
-import java.math.BigDecimal;
-
-@Getter
-public class CurrencyRequestDto {
-    private String currencyName;
-    private BigDecimal exchangeRate;
-    private String symbol;
-
-    public Currency toEntity() {
-        return new Currency(
-                this.currencyName,
-                this.exchangeRate,
-                this.symbol
-        );
-    }
-}
