--- src/main/java/com/sparta/currency_user/currency/service/CurrencyService.java
@@ -1,36 +1,0 @@
-package com.sparta.currency_user.currency.service;
-
-import com.sparta.currency_user.currency.dto.CurrencyRequestDto;
-import com.sparta.currency_user.currency.dto.CurrencyResponseDto;
-import com.sparta.currency_user.currency.entity.Currency;
-import com.sparta.currency_user.currency.repository.CurrencyRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class CurrencyService {
-
-    private final CurrencyRepository currencyRepository;
-
-    public CurrencyResponseDto findById(Long id) {
-        return new CurrencyResponseDto(findCurrencyById(id));
-    }
-
-    public Currency findCurrencyById(Long id) {
-        return currencyRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("통화를 찾을 수 없습니다."));
-    }
-
-    public List<CurrencyResponseDto> findAll() {
-        return currencyRepository.findAll().stream().map(CurrencyResponseDto::toDto).toList();
-    }
-
-    @Transactional
-    public CurrencyResponseDto save(CurrencyRequestDto currencyRequestDto) {
-        Currency savedCurrency = currencyRepository.save(currencyRequestDto.toEntity());
-        return new CurrencyResponseDto(savedCurrency);
-    }
-}
