+++ src/main/java/com/sparta/currency_user/domain/currency/entity/Currency.java
@@ -1,0 +1,28 @@
+package com.sparta.currency_user.domain.currency.entity;
+
+import com.sparta.currency_user.domain.base.entity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.math.BigDecimal;
+
+@Entity
+@Getter
+@Table(name = "currency")
+public class Currency extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String currencyName;
+    private BigDecimal exchangeRate;
+    private String symbol;
+
+    public Currency(String currencyName, BigDecimal exchangeRate, String symbol) {
+        this.currencyName = currencyName;
+        this.exchangeRate = exchangeRate;
+        this.symbol = symbol;
+    }
+
+    public Currency() {}
+}
