+++ src/main/java/com/sparta/currency_user/domain/currency/service/CurrencyService.java
@@ -1,0 +1,36 @@
+package com.sparta.currency_user.domain.currency.service;
+
+import com.sparta.currency_user.domain.currency.dto.CurrencyRequestDto;
+import com.sparta.currency_user.domain.currency.dto.CurrencyResponseDto;
+import com.sparta.currency_user.domain.currency.entity.Currency;
+import com.sparta.currency_user.domain.currency.repository.CurrencyRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class CurrencyService {
+
+    private final CurrencyRepository currencyRepository;
+
+    public CurrencyResponseDto findById(Long id) {
+        return new CurrencyResponseDto(findCurrencyById(id));
+    }
+
+    public Currency findCurrencyById(Long id) {
+        return currencyRepository.findByIdOrElseThrow(id);
+    }
+
+    public List<CurrencyResponseDto> findAll() {
+        return currencyRepository.findAll().stream().map(CurrencyResponseDto::toDto).toList();
+    }
+
+    @Transactional
+    public CurrencyResponseDto save(CurrencyRequestDto currencyRequestDto) {
+        Currency savedCurrency = currencyRepository.save(currencyRequestDto.toEntity());
+        return new CurrencyResponseDto(savedCurrency);
+    }
+}
