+++ src/main/java/com/sparta/currency_user/domain/exchange/enums/ExchangeStatus.java
@@ -1,0 +1,5 @@
+package com.sparta.currency_user.domain.exchange.enums;
+
+public enum ExchangeStatus {
+    NORMAL, CANCELLED
+}
