+++ src/main/java/com/sparta/currency_user/domain/user/dto/UserRequestDto.java
@@ -1,0 +1,26 @@
+package com.sparta.currency_user.domain.user.dto;
+
+import com.sparta.currency_user.domain.user.entity.User;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+import lombok.Getter;
+
+@Getter
+public class UserRequestDto {
+
+    @NotBlank
+    @Size(min = 2, max = 10)
+    private String name;
+
+    @NotBlank
+    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+.[A-Za-z]{2,6}$")
+    private String email;
+
+    public User toEntity() {
+        return new User(
+                this.name,
+                this.email
+        );
+    }
+}
