+++ src/main/java/com/sparta/currency_user/domain/user/service/UserService.java
@@ -1,0 +1,46 @@
+package com.sparta.currency_user.domain.user.service;
+
+import com.sparta.currency_user.domain.user.dto.UserRequestDto;
+import com.sparta.currency_user.domain.user.dto.UserResponseDto;
+import com.sparta.currency_user.domain.user.entity.User;
+import com.sparta.currency_user.domain.user.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+public class UserService {
+    private final UserRepository userRepository;
+
+    @Autowired
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Transactional
+    public UserResponseDto save(UserRequestDto userRequestDto) {
+        User savedUser = userRepository.save(userRequestDto.toEntity());
+        return new UserResponseDto(savedUser);
+    }
+
+    public List<UserResponseDto> findAll() {
+        return userRepository.findAll().stream().map(UserResponseDto::toDto).toList();
+    }
+
+    public UserResponseDto findById(Long id) {
+        return new UserResponseDto(findUserById(id));
+    }
+
+    public User findUserById(Long id) {
+        return userRepository.findByIdOrElseThrow(id);
+    }
+
+    @Transactional
+    public void deleteUserById(Long id) {
+        this.findUserById(id);
+        userRepository.deleteById(id);
+    }
+
+}
