--- src/main/java/com/sparta/currency_user/exchange/enums/ExchangeStatus.java
@@ -1,5 +1,0 @@
-package com.sparta.currency_user.exchange.enums;
-
-public enum ExchangeStatus {
-    NORMAL, CANCELLED
-}
