+++ src/main/java/com/sparta/currency_user/global/error/NotFoundByIdException.java
@@ -1,0 +1,11 @@
+package com.sparta.currency_user.global.error;
+
+import lombok.Getter;
+
+@Getter
+public class NotFoundByIdException extends BusinessException {
+
+    public NotFoundByIdException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
