--- src/main/java/com/sparta/currency_user/user/controller/UserController.java
@@ -1,39 +1,0 @@
-package com.sparta.currency_user.user.controller;
-
-import com.sparta.currency_user.user.dto.UserRequestDto;
-import com.sparta.currency_user.user.dto.UserResponseDto;
-import com.sparta.currency_user.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/users")
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-
-    @GetMapping
-    public ResponseEntity<List<UserResponseDto>> findUsers() {
-        return ResponseEntity.ok().body(userService.findAll());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<UserResponseDto> findUser(@PathVariable Long id) {
-        return ResponseEntity.ok().body(userService.findById(id));
-    }
-
-    @PostMapping
-    public ResponseEntity<UserResponseDto> createUser(@RequestBody UserRequestDto userRequestDto) {
-        return ResponseEntity.ok().body(userService.save(userRequestDto));
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
-        userService.deleteUserById(id);
-        return ResponseEntity.ok().body("정상적으로 삭제되었습니다.");
-    }
-}
