--- src/main/java/com/sparta/currency_user/user/dto/UserResponseDto.java
@@ -1,32 +1,0 @@
-package com.sparta.currency_user.user.dto;
-
-import com.sparta.currency_user.user.entity.User;
-import lombok.Getter;
-
-@Getter
-public class UserResponseDto {
-    private final Long id;
-
-    private final String name;
-    private final String email;
-
-    public UserResponseDto(User user) {
-        this.id = user.getId();
-        this.name = user.getName();
-        this.email = user.getEmail();
-    }
-
-    public UserResponseDto(Long id, String name, String email) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-    }
-
-    public static UserResponseDto toDto(User user) {
-        return new UserResponseDto(
-                user.getId(),
-                user.getName(),
-                user.getEmail()
-        );
-    }
-}
