--- src/main/java/com/sparta/currency_user/user/entity/User.java
@@ -1,31 +1,0 @@
-package com.sparta.currency_user.user.entity;
-
-import com.sparta.currency_user.base.entity.BaseEntity;
-import com.sparta.currency_user.exchange.entity.Exchange;
-import jakarta.persistence.*;
-import lombok.Getter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Entity
-@Table(name = "user")
-public class User extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-    private String email;
-
-    @OneToMany(mappedBy = "user", orphanRemoval = true, cascade = CascadeType.REMOVE)
-    private List<Exchange> exchanges = new ArrayList<>();
-
-    public User(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
-
-    public User() {}
-}
